Plugins: categorize ldrmodules as a malware plugin
